data_cafe/requirements.txt
playwright
requests
fastapi
uvicorn
supabase
python-dotenv
opencv-python-headless
numpy
imagehash
Pillow
